refactor: Move remaining error/logging code to `core-interfaces`
