+++ packages/utils/telemetry-utils/package.json
@@ -71,8 +71,7 @@
 		"@fluidframework/core-interfaces": "workspace:~",
 		"@fluidframework/core-utils": "workspace:~",
 		"debug": "^4.1.1",
-		"events": "^3.1.0",
-		"uuid": "^8.3.1"
+		"events": "^3.1.0"
 	},
 	"devDependencies": {
 		"@fluid-tools/build-cli": "^0.22.0",
@@ -99,7 +98,8 @@
 		"prettier": "~2.6.2",
 		"rimraf": "^4.4.0",
 		"sinon": "^7.4.2",
-		"typescript": "~4.5.5"
+		"typescript": "~4.5.5",
+		"uuid": "^8.3.1"
 	},
 	"typeValidation": {
 		"broken": {}
